backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
pypdf
passlib[bcrypt]
python-multipart
psycopg2-binary